ValirdarCamposVacios

+++ BazarRoxana/Categorias.vb
@@ -4,7 +4,7 @@ Public Class Categorias
     Private Sub Categorias_Load(sender As Object, e As EventArgs) Handles MyBase.Load
 
         Dim conec As New SqlClient.SqlConnection
-        conec.ConnectionString = "Data Source=DANIELRUEDA\LOCALHOST;Initial Catalog=BazarRoxana;Integrated Security=True"
+        conec.ConnectionString = "Data Source=AMAYA;Initial Catalog=BazarRoxana;Integrated Security=True"
         conec.Open()
 
         Dim DatosCat As New DataTable 'tabla temporal que recoge los datos de la consulta
@@ -20,80 +20,76 @@ Public Class Categorias
     Private Sub btnGuardar_Click(sender As Object, e As EventArgs) Handles btnGuardar.Click
 
         Dim conec As New SqlClient.SqlConnection
-        conec.ConnectionString = "Data Source=DANIELRUEDA\LOCALHOST;Initial Catalog=BazarRoxana;Integrated Security=True"
+        conec.ConnectionString = "Data Source=AMAYA;Initial Catalog=BazarRoxana;Integrated Security=True"
         conec.Open()
-
-        Dim consultaGuardar As String = "insert into Categoria(CodCateg, NombCateg) values(@CodCateg, @NombCateg)"
-
-        Dim ejecutar As New SqlCommand(consultaGuardar, conec)
-
-        ejecutar.Parameters.AddWithValue("@CodCateg", Val(txCodCat.Text))
-        ejecutar.Parameters.AddWithValue("@NombCateg", (txtNombCat.Text))
-
-
-        ejecutar.ExecuteNonQuery()
-
         Dim DatosCat As New DataTable 'tabla temporal que recoge los datos de la consulta
         Using adaptador As New SqlDataAdapter("select * from Categoria", conec)
             adaptador.Fill(DatosCat)
         End Using 'intermediario entre la base de datos y DATOSusuario para poder ingresar a datatable
-
         DGV.DataSource = DatosCat
+        If txCodCat.Text = "" Or txtNombCat.Text = "" Then
+            MsgBox("Hay campos vacios")
+        Else
+            Dim consultaGuardar As String = "insert into Categoria(CodCateg, NombCateg) values(@CodCateg, @NombCateg)"
+            Dim ejecutar As New SqlCommand(consultaGuardar, conec)
+            ejecutar.Parameters.AddWithValue("@CodCateg", Val(txCodCat.Text))
+            ejecutar.Parameters.AddWithValue("@NombCateg", (txtNombCat.Text))
+            ejecutar.ExecuteNonQuery()
 
+        End If
     End Sub
 
     Private Sub btnActualizar_Click(sender As Object, e As EventArgs) Handles btnActualizar.Click
 
         Dim conec As New SqlClient.SqlConnection
-        conec.ConnectionString = "Data Source=DANIELRUEDA\LOCALHOST;Initial Catalog=BazarRoxana;Integrated Security=True"
+        conec.ConnectionString = "Data Source=AMAYA;Initial Catalog=BazarRoxana;Integrated Security=True"
         conec.Open()
-
-        Dim consultaAct As String = "update Categoria set NombCateg=@NombCateg where CodCateg= @CodCateg"
-
-        Dim ejecutar As New SqlCommand(consultaAct, conec)
-
-        ejecutar.Parameters.AddWithValue("@CodCateg", Val(txCodCat.Text))
-        ejecutar.Parameters.AddWithValue("@NombCateg", (txtNombCat.Text))
-
-
-        ejecutar.ExecuteNonQuery()
-
         Dim DatosCat As New DataTable 'tabla temporal que recoge los datos de la consulta
         Using adaptador As New SqlDataAdapter("select * from Categoria", conec)
             adaptador.Fill(DatosCat)
         End Using 'intermediario entre la base de datos y DATOSusuario para poder ingresar a datatable
 
         DGV.DataSource = DatosCat
+        If txCodCat.Text = "" Or txtNombCat.Text = "" Then
+            MsgBox("Hay campos vacios")
+        Else
 
+            Dim consultaAct As String = "update Categoria set NombCateg=@NombCateg where CodCateg= @CodCateg"
+            Dim ejecutar As New SqlCommand(consultaAct, conec)
+            ejecutar.Parameters.AddWithValue("@CodCateg", Val(txCodCat.Text))
+            ejecutar.Parameters.AddWithValue("@NombCateg", (txtNombCat.Text))
+            ejecutar.ExecuteNonQuery()
+
+        End If
     End Sub
 
     Private Sub btnEliminar_Click(sender As Object, e As EventArgs) Handles btnEliminar.Click
 
         Dim conec As New SqlClient.SqlConnection
-        conec.ConnectionString = "Data Source=DANIELRUEDA\LOCALHOST;Initial Catalog=BazarRoxana;Integrated Security=True"
+        conec.ConnectionString = "Data Source=AMAYA;Initial Catalog=BazarRoxana;Integrated Security=True"
         conec.Open()
-
-        Dim consultaElim As String = "delete from Categoria  where CodCateg= @CodCateg"
-
-        Dim ejecutar As New SqlCommand(consultaElim, conec)
-
-        ejecutar.Parameters.AddWithValue("@CodCateg", Val(txCodCat.Text))
-
-        ejecutar.ExecuteNonQuery()
 
         Dim DatosCat As New DataTable 'tabla temporal que recoge los datos de la consulta
         Using adaptador As New SqlDataAdapter("select * from Categoria", conec)
             adaptador.Fill(DatosCat)
         End Using 'intermediario entre la base de datos y DATOSusuario para poder ingresar a datatable
-
         DGV.DataSource = DatosCat
+        If txCodCat.Text = "" Then
+            MsgBox("Hay campos vacios")
+        Else
 
+            Dim consultaElim As String = "delete from Categoria  where CodCateg= @CodCateg"
+            Dim ejecutar As New SqlCommand(consultaElim, conec)
+            ejecutar.Parameters.AddWithValue("@CodCateg", Val(txCodCat.Text))
+            ejecutar.ExecuteNonQuery()
+
+        End If
     End Sub
 
     Private Sub btnBuscar_Click(sender As Object, e As EventArgs) Handles btnBuscar.Click
 
         Dim conec As New SqlClient.SqlConnection
-        conec.ConnectionString = "Data Source=DANIELRUEDA\LOCALHOST;Initial Catalog=BazarRoxana;Integrated Security=True"
+        conec.ConnectionString = "Data Source=AMAYA;Initial Catalog=BazarRoxana;Integrated Security=True"
         conec.Open()
 
         Dim busqueda As Integer
@@ -115,7 +111,7 @@ Public Class Categorias
     Private Sub btnActTabla_Click(sender As Object, e As EventArgs) Handles btnActTabla.Click
 
         Dim conec As New SqlClient.SqlConnection
-        conec.ConnectionString = "Data Source=DANIELRUEDA\LOCALHOST;Initial Catalog=BazarRoxana;Integrated Security=True"
+        conec.ConnectionString = "Data Source=AMAYA;Initial Catalog=BazarRoxana;Integrated Security=True"
         conec.Open()
 
         Dim DatosCat As New DataTable 'tabla temporal que recoge los datos de la consulta
