+++ BazarRoxana/Empleado.vb
@@ -4,7 +4,7 @@ Public Class Empleado
     Private Sub Empleado_Load(sender As Object, e As EventArgs) Handles MyBase.Load
 
         Dim conec As New SqlClient.SqlConnection
-        conec.ConnectionString = "Data Source=CAROLINA10\CAROLINA;Initial Catalog=BazarRoxana;Integrated Security=True"
+        conec.ConnectionString = "Data Source=AMAYA;Initial Catalog=BazarRoxana;Integrated Security=True"
         conec.Open()
         'CAROLINA10\CAROLINA
         'AMAYA
@@ -21,82 +21,95 @@ Public Class Empleado
 
     Private Sub btnGuardar_Click(sender As Object, e As EventArgs) Handles btnGuardar.Click
         Dim conec As New SqlClient.SqlConnection
-        conec.ConnectionString = "Data Source=CAROLINA10\CAROLINA;Initial Catalog=BazarRoxana;Integrated Security=True"
+        conec.ConnectionString = "Data Source=AMAYA;Initial Catalog=BazarRoxana;Integrated Security=True"
         conec.Open()
-        Dim Cat As Integer
-
-        If CbxNivel.SelectedItem = "Gerente" Then
-            Cat = 1
-        Else
-            Cat = 2
-        End If
-
-        Dim consultaGuardar As String = "insert into Empleados(CodEmple, NombEmple, Contraseña,NivelEmple,EstadoEmple) values(@CodEmple, @NombEmple,@Contraseña,@NivelEmple,1)"
-        Dim ejecutar As New SqlCommand(consultaGuardar, conec)
-        ejecutar.Parameters.AddWithValue("@CodEmple", Val(TxtCodigoEmpleado.Text))
-        ejecutar.Parameters.AddWithValue("@NombEmple", (txtNombreEmpleado.Text))
-        ejecutar.Parameters.AddWithValue("@Contraseña", (TxtContraseña.Text))
-        ejecutar.Parameters.AddWithValue("@NivelEmple", (Cat))
 
 
-        ejecutar.ExecuteNonQuery()
         Dim DatosEmp As New DataTable 'tabla temporal que recoge los datos de la consulta
         Using adaptador As New SqlDataAdapter("select * from Empleados", conec)
             adaptador.Fill(DatosEmp)
         End Using 'intermediario entre la base de datos y DATOSusuario para poder ingresar a datatable
 
         DGV.DataSource = DatosEmp
+        If TxtCodigoEmpleado.Text = "" Or txtNombreEmpleado.Text = "" Or TxtContraseña.Text = "" Then
+            MsgBox("Hay campos vacios")
+        Else
+
+            Dim Cat As Integer
+
+            If CbxNivel.SelectedItem = "Gerente" Then
+                Cat = 1
+            Else
+                Cat = 2
+            End If
+
+            Dim consultaGuardar As String = "insert into Empleados(CodEmple, NombEmple, Contraseña,NivelEmple,EstadoEmple) values(@CodEmple, @NombEmple,@Contraseña,@NivelEmple,1)"
+            Dim ejecutar As New SqlCommand(consultaGuardar, conec)
+            ejecutar.Parameters.AddWithValue("@CodEmple", Val(TxtCodigoEmpleado.Text))
+            ejecutar.Parameters.AddWithValue("@NombEmple", (txtNombreEmpleado.Text))
+            ejecutar.Parameters.AddWithValue("@Contraseña", (TxtContraseña.Text))
+            ejecutar.Parameters.AddWithValue("@NivelEmple", (Cat))
+            ejecutar.ExecuteNonQuery()
+
+        End If
+
     End Sub
 
     Private Sub btnActualizar_Click(sender As Object, e As EventArgs) Handles btnActualizar.Click
         Dim conec As New SqlClient.SqlConnection
-        conec.ConnectionString = "Data Source=CAROLINA10\CAROLINA;Initial Catalog=BazarRoxana;Integrated Security=True"
+        conec.ConnectionString = "Data Source=AMAYA;Initial Catalog=BazarRoxana;Integrated Security=True"
         conec.Open()
-        Dim Cat As Integer
-
-        If CbxNivel.SelectedItem = "Gerente" Then
-            Cat = 1
-        Else
-            Cat = 2
-        End If
-
-        Dim consultaAct As String = "update Empleados set NombEmple=@NombEmple, NivelEmple=@NivelEmple where CodEmple= @CodEmple"
-        Dim ejecutar As New SqlCommand(consultaAct, conec)
-        ejecutar.Parameters.AddWithValue("@CodEmple", Val(TxtCodigoEmpleado.Text))
-        ejecutar.Parameters.AddWithValue("@NombEmple", (txtNombreEmpleado.Text))
-        ejecutar.Parameters.AddWithValue("@NivelEmple", (Cat))
-
-
-        ejecutar.ExecuteNonQuery()
         Dim DatosEmp As New DataTable 'tabla temporal que recoge los datos de la consulta
         Using adaptador As New SqlDataAdapter("select * from Empleados", conec)
             adaptador.Fill(DatosEmp)
         End Using 'intermediario entre la base de datos y DATOSusuario para poder ingresar a datatable
-
         DGV.DataSource = DatosEmp
+        If TxtCodigoEmpleado.Text = "" Or txtNombreEmpleado.Text = "" Or TxtContraseña.Text = "" Then
+            MsgBox("Hay campos vacios")
+        Else
+            Dim Cat As Integer
+
+            If CbxNivel.SelectedItem = "Gerente" Then
+                Cat = 1
+            Else
+                Cat = 2
+            End If
+
+            Dim consultaAct As String = "update Empleados set NombEmple=@NombEmple, NivelEmple=@NivelEmple where CodEmple= @CodEmple"
+            Dim ejecutar As New SqlCommand(consultaAct, conec)
+            ejecutar.Parameters.AddWithValue("@CodEmple", Val(TxtCodigoEmpleado.Text))
+            ejecutar.Parameters.AddWithValue("@NombEmple", (txtNombreEmpleado.Text))
+            ejecutar.Parameters.AddWithValue("@NivelEmple", (Cat))
+            ejecutar.ExecuteNonQuery()
+
+        End If
     End Sub
 
     Private Sub btnEliminar_Click(sender As Object, e As EventArgs) Handles btnEliminar.Click
         Dim conec As New SqlClient.SqlConnection
-        conec.ConnectionString = "Data Source=CAROLINA10\CAROLINA;Initial Catalog=BazarRoxana;Integrated Security=True"
+        conec.ConnectionString = "Data Source=AMAYA;Initial Catalog=BazarRoxana;Integrated Security=True"
         conec.Open()
 
-        Dim consultaElim As String = "delete from Empleados  where CodEmple= @CodEmple"
-        Dim ejecutar As New SqlCommand(consultaElim, conec)
-        ejecutar.Parameters.AddWithValue("@CodEmple", Val(TxtCodigoEmpleado.Text))
-
-        ejecutar.ExecuteNonQuery()
         Dim DatosEmp As New DataTable 'tabla temporal que recoge los datos de la consulta
         Using adaptador As New SqlDataAdapter("select * from Empleados", conec)
             adaptador.Fill(DatosEmp)
         End Using 'intermediario entre la base de datos y DATOSusuario para poder ingresar a datatable
 
         DGV.DataSource = DatosEmp
+        If TxtCodigoEmpleado.Text = "" Then
+            MsgBox("Hay campos vacios")
+        Else
+            Dim consultaElim As String = "delete from Empleados  where CodEmple= @CodEmple"
+            Dim ejecutar As New SqlCommand(consultaElim, conec)
+            ejecutar.Parameters.AddWithValue("@CodEmple", Val(TxtCodigoEmpleado.Text))
+
+            ejecutar.ExecuteNonQuery()
+        End If
     End Sub
 
     Private Sub btnBuscar_Click(sender As Object, e As EventArgs) Handles btnBuscar.Click
         Dim conec As New SqlClient.SqlConnection
-        conec.ConnectionString = "Data Source=CAROLINA10\CAROLINA;Initial Catalog=BazarRoxana;Integrated Security=True"
+        conec.ConnectionString = "Data Source=AMAYA;Initial Catalog=BazarRoxana;Integrated Security=True"
         conec.Open()
 
         Dim busqueda As Integer
@@ -128,7 +141,7 @@ Public Class Empleado
 
     Private Sub btnActTabla_Click(sender As Object, e As EventArgs) Handles btnActTabla.Click
         Dim conec As New SqlClient.SqlConnection
-        conec.ConnectionString = "Data Source=CAROLINA10\CAROLINA;Initial Catalog=BazarRoxana;Integrated Security=True"
+        conec.ConnectionString = "Data Source=AMAYA;Initial Catalog=BazarRoxana;Integrated Security=True"
         conec.Open()
 
         Dim DatosEmp As New DataTable 'tabla temporal que recoge los datos de la consulta
@@ -173,5 +186,7 @@ Public Class Empleado
 
     Private Sub cbxMostrarContra_CheckedChanged(sender As Object, e As EventArgs) Handles cbxMostrarContra.CheckedChanged
         TxtContraseña.UseSystemPasswordChar = Not cbxMostrarContra.Checked
+
     End Sub
+
 End Class
